Cleanup font debug and unnecessary guides prefs cast. Remove column/ frame/link breaks when requested from paragraph styles
git-svn-id: svn://scribus.net/trunk/Scribus@26062 11d20701-8431-0410-a711-e3c959e3b870

+++ scribus/ui/prefs_guides.cpp
@@ -304,7 +304,7 @@ void Prefs_Guides::moveUp()
 void Prefs_Guides::moveDown()
 {
 	int curr = guidePlacementListBox->currentRow();
-	if (curr == static_cast<int>(guidePlacementListBox->count()) - 1)
+	if (curr == guidePlacementListBox->count() - 1)
 		return;
 	QListWidgetItem *it = guidePlacementListBox->takeItem(curr);
 	guidePlacementListBox->insertItem(curr + 1, it);
